System connection

+++ raspberry/lib/communication/communication.h
@@ -34,13 +34,11 @@
 /*
  *	Create a new socket.
  */
-int createServerSocket(int port, int numClient);
-
+int createServerSocket(int port, uint8_t numClient);
 /*
  *	Create a new client socket.
  */
 int createClientSocket(const char *ip, int port);
-
 /*
  *	Wait for reading a data package.
  */
